Avoid whitespace changes from data sources
There is a mismatch between the resource that is built by
the build and overlay kustomization data sources and the
resource information returned by kubernetes.

This seems to show up after first import only, but it's worth
being consistent

Because we use a yaml.Encoder to provide the manifest to
send to kubernetes, it ends in a trailing newline (and this
is reflected in the test suite). However, the current state
is pulled from lastAppliedConfig which doesn't have that
whitespace, and so the resources appear to mismatch and need
updating.

Adding a trailing new line to the existing state before the
compare seems to minimise required changes.